Sync script that works for different folks.

+++ sync_with_astm.sh
@@ -1,8 +1,15 @@
 # Sync code to astm and notebooks from astm.
-REMOTE_DIR=/opt/astro/devel/yuping/distributed-pipeline/
-LOCAL_DIR=/home/yuping/radio_projects/lwa/distributed-pipeline/
-REMOTE_HOST=astmhead
-REMOTE_USER=yuping
+set -e
+
+if [ -z "$ASTM_USER_NAME" ]; then
+  echo 'Environmental variable ASTM_USER_NAME is not set.'
+  exit 1
+fi
+
+REMOTE_USER=$ASTM_USER_NAME
+REMOTE_DIR=/opt/astro/devel/${REMOTE_USER}/distributed-pipeline/
+LOCAL_DIR=`pwd`
+REMOTE_HOST=astm13.lwa.ovro.caltech.edu
 
 rsync -aP --delete --exclude 'notebooks' --exclude '.pytest_cache' --exclude '.venv' \
     --exclude '*__pycache__*' --exclude 'ovro_lwa_orca.egg-info' \
